[diffusion] Refine opt-in conditioning expansion
